ROS-210: fix flashing of T41
Update FLASH_RESERVE for T41 to account for the sectors used for SEEPROM
emulation plus the recovery sector,, which according to this thread
(https://forum.pjrc.com/threads/43165-Over-the-Air-firmware-updates-changes-for-flashing-Teensy-3-5-amp-3-6?p=317345&viewfull=1#post317345)
is 64 sectors. This seems to match what I was seeing as the WeedbotMCU is
storing some data in the emulated EEPROM and I think this is what the
flasher was stumbling across.